ensure that pybuild knows where to install
With only one binary package, pybuild always put files in the right
place.

now that we're producing two binary packages, pybuild guesses the name
of the package based on the PYBUILD_NAME variable, and comes up with
python3-impass.  it installs things there, which doesn't actually do
anything for the rest of the build.

Use PYBUILD_DESTDIR to override its mistaken choice.