gui: convert to GTK 4, drop GtkBuilder XML
GTK 4 does a better rendering job on sway.  In particular, it doesn't
cause weird flickering on secondary displays due to
GtkEntryCompletion.

However, GtkEntryCompletion is itself also deprecated, so we drop
GtkEntryCompletion entirely in favor of a GtkPopup.

There were a number of substantive changes that needed to happen due
to changes from GTK 3 to GTK 4 (see
https://docs.gtk.org/gtk4/migrating-3to4.html )

- drop gtk_main, rely on GtkApplication instead
- drop deprecated GtkEntryCompletion, in favor of a self-made GtkPopup
- drop GtkMenu in favor of GtkPopupMenu
- use Actions where it makes sense to do so

Outstanding things to do to reach a satisfying GTK4 state:

- restore context menu additions ("populate-popup")
- remove unnecessary ctxwarning Box (ctxwarninglabel should be fine)
- use a single GtkPopupMenu for custom/delete
- repace custom GtkPopup with GtkListView, populated from a ListModel
  and a Factory